Push run logic into settings.
This is probably just a step towards a better overall shape, but it
seemed like the run function in lib was doing too much. It also seems
like settings is doing too much, now, so it might make sense to seek a
better structure for that. Still, this puts us in a spot where we can
refactor the responsibility of individual components into their
respective packages.