Add WASM client.
This one took a bit of indirection, and it influenced project structure.
Hopefully for the better?

The original design of the service kit setup meant that different
compile targets would conflict with each other, so this commit
introduces a diamond-shaped structure instead. The main crate has been
turned into a workspace, and divided into the following crates:

- Service kit core, which is the original setup more or less intact.
- Service kit, which leverages it and provides the command binary.
- Service kit web, which has the wasm client inside.
- xtask, which manages the wasm bindgen command.

This setup allows us to compile wasm, and pop it into the right
directory, and provides a basic proof of concept for its work. To run
the build, use this command:

```sh
cargo xtask web-build
```

That will emit the wasm files into the example JS dist folder for the
server.

Some stuff left on the table that leaves us in an incomplete state:

- The service kit web should depend on the types from service kit core,
  somehow, but right now that's not the case. That means we have two
  clients, a rust client and a web client.
- The actual shipping of the client isn't being done at all and I'm not
  sure if cargo-dist will help us with that part. So we'll have to
  invent it.