Add sqlite file storage.
This is a very minimal implementation of sqlite file storage. The spirit
of it is this: instead of storing files on the host server running
service kits, we generally want to bundle those files up and use sqlite
as a custom file storage format instead. This gives us the ability to
make a single service a lot more portable later if we like, eliminating
an entire category of issues revolving around file management.

Unfortunately this leaves a lot on the table which I'd hoped to add
along with storage, and so I'm leaving a note here that I wish I'd been
able to do these things, too:

- Create a single "storage" metaphor that handled memory and file
  representations both.
- Create a graceful shutdown that copies memory to a separate cache db,
  and rehydrated on next boot up.
- Allow consumers to access sqlite pools and do their own stuff with
  them.

I say "unfortunately" not because those are must-haves. I think this
might be a better architecture since it'll encourage me to represent
different sqlite lifecycles as different types. It's unfortunate because
I spent two weeks trying to debug sqlx issues and part of that I thought
was my fault. Delays suck!