Add protocol buffers.
This adds prost and protobuf support to the crates, which are meant to
let us design clients and services around the same type strata. A
vestigial service generator is bundled, which has no complex support for
anything like tower services or things of that nature, but some notes
are left on how to get there if we want to later.