Create protocol crate and leverage in clients.
I realized I hadn't actually leveraged the protobuf stuff in the wasm
layer. Investigating ways to do that, I realized that a diamond crate
structure would be necessary if I didn't want to employ a bunch of
config flags to create custom compilations. So that's what this commit
does: it redistributes the protocol from its own crate to the two
consumer core and web crates.