Carve up settings modules.
While things were being spiked out, settings sprung into life pretty
quickly, and so I chose to just keep them in one module. That got
unwieldy fast, and so this commit organizes our settings somewhat. It
may not stick like this but it should be a little easier to read later,
at least.